Register components for GCS, HDFS FileSystem during HiveConnector registration
