+++ velox/connectors/hive/storage_adapters/gcs/CMakeLists.txt
@@ -14,17 +14,21 @@
 
 # for generated headers
 
-add_library(velox_gcs GCSFileSystem.cpp GCSUtil.cpp)
-target_link_libraries(velox_gcs Folly::folly google-cloud-cpp::storage)
+add_library(velox_gcs RegisterGCSFileSystem.cpp)
 
-if(${VELOX_BUILD_TESTING})
-  add_subdirectory(tests)
-endif()
+if(VELOX_ENABLE_GCS)
+  target_sources(velox_gcs PRIVATE GCSFileSystem.cpp GCSUtil.cpp)
+  target_link_libraries(velox_gcs Folly::folly google-cloud-cpp::storage)
 
-if(${VELOX_ENABLE_BENCHMARKS})
-  add_subdirectory(benchmark)
-endif()
+  if(${VELOX_BUILD_TESTING})
+    add_subdirectory(tests)
+  endif()
 
-if(${VELOX_ENABLE_EXAMPLES})
-  add_subdirectory(examples)
+  if(${VELOX_ENABLE_BENCHMARKS})
+    add_subdirectory(benchmark)
+  endif()
+
+  if(${VELOX_ENABLE_EXAMPLES})
+    add_subdirectory(examples)
+  endif()
 endif()
